Updated UI components and added new chat features
Updated the solution file to use a different project type GUID and added a new JavaScript file to the App.razor file. Cleared the ChatView.razor file of its previous content and updated the binding for the MudSwitch component in MainLayout.razor. Added a new MudButton to NavMenu.razor and updated Home.razor to include a ChatView component. Added a new property to the UserService.cs class and added new components with their respective code. Also, added a new JavaScript function to prevent the default action when the Enter key is pressed.